#615: moved cplane files to internal package
