+++ internal/controlplane/config_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/NVIDIA/k8s-test-infra/pkg/controlplane"
+	"github.com/NVIDIA/k8s-test-infra/internal/controlplane"
 	"github.com/stretchr/testify/require"
 )
